added first part of the activity calculation algo

+++ src/main/scala/helix/lib/contracts.scala
@@ -27,6 +27,6 @@ trait Scoring {
   protected def scoring: ScoringStrategy
   
   trait ScoringStrategy {
-    def calculateProjectActivityScore(project: Project): Int
+    def calculateProjectActivityScore(project: Project): BigInt
   }
 }
